third_party/WebKit/LayoutTests/media/video-src-set.html

<!DOCTYPE html>
<title>Test that setting "src" attribute triggers load.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    var mediaFile = findMediaFile("video", "content/test");

    video.onloadstart = t.step_func_done(function() {
        var url = video.currentSrc;
        assert_equals(url.substr(url.lastIndexOf("/media/")+7), mediaFile);
    });

    video.src = mediaFile;
});
</script>
